Completed Basic GUI
The program now features a basic GUI that allows players to select a
cell and move the piece in that cell to another selected cell.

ChessBoard has been given a new function, getPieceByPos(), that is used
to relate the coordinates chosen in the GUI to those in the 2-D array.